Monitoring - Display data of cursor position

+++ apphub-lib/apphub-lib-monitoring/src/main/java/com/github/saphyra/apphub/lib/monitoring/MemoryStatusReporter.java
@@ -18,15 +18,15 @@ public class MemoryStatusReporter {
     private static final String KEY_AVAILABLE_MEMORY = "availableMemory";
     private static final String KEY_ALLOCATED_MEMORY = "allocatedMemory";
     private static final String KEY_USED_MEMORY = "usedMemory";
-    private static final long BYTES_IN_MEGABYTE = 1024L * 1024L;
+    private static final double BYTES_IN_MEGABYTE = 1024d * 1024d;
 
     private final MetricRegistry metricRegistry;
 
     public void reportMemoryStatus() {
         Runtime runtime = Runtime.getRuntime();
-        long totalMemory = runtime.totalMemory() / BYTES_IN_MEGABYTE;
-        long freeMemory = runtime.freeMemory() / BYTES_IN_MEGABYTE;
-        long maxMemory = runtime.maxMemory() / BYTES_IN_MEGABYTE;
+        double totalMemory = runtime.totalMemory() / BYTES_IN_MEGABYTE;
+        double freeMemory = runtime.freeMemory() / BYTES_IN_MEGABYTE;
+        double maxMemory = runtime.maxMemory() / BYTES_IN_MEGABYTE;
 
         metricRegistry.reportMetric(
             Feature.MEMORY_MONITORING,
@@ -34,17 +34,17 @@ public class MemoryStatusReporter {
             List.of(
                 MetricPropertyModel.builder()
                     .key(KEY_AVAILABLE_MEMORY)
-                    .value((double) maxMemory)
+                    .value(maxMemory)
                     .aggregationStrategy(AggregationStrategy.MAX)
                     .build(),
                 MetricPropertyModel.builder()
                     .key(KEY_ALLOCATED_MEMORY)
-                    .value((double) totalMemory)
+                    .value(totalMemory)
                     .aggregationStrategy(AggregationStrategy.AVERAGE)
                     .build(),
                 MetricPropertyModel.builder()
                     .key(KEY_USED_MEMORY)
-                    .value((double) totalMemory - freeMemory)
+                    .value(totalMemory - freeMemory)
                     .aggregationStrategy(AggregationStrategy.AVERAGE)
                     .build()
             )
